+++ genai_otel/instrumentors/__init__.py
@@ -47,6 +47,7 @@ from .replicate_instrumentor import ReplicateInstrumentor
 from .sambanova_instrumentor import SambaNovaInstrumentor
 from .sarvam_instrumentor import SarvamAIInstrumentor
 from .sentence_transformers_instrumentor import SentenceTransformersInstrumentor
+from .strands_harness_instrumentor import StrandsHarnessInstrumentor
 from .togetherai_instrumentor import TogetherAIInstrumentor
 from .typesafe_instrumentor import TypeSafeInstrumentor
 from .vertexai_instrumentor import VertexAIInstrumentor
@@ -94,4 +95,5 @@ __all__ = [
     "SarvamAIInstrumentor",
     "SentenceTransformersInstrumentor",
     "TypeSafeInstrumentor",
+    "StrandsHarnessInstrumentor",
 ]